+++ pyo3-ffi/src/iterobject.rs
@@ -9,7 +9,7 @@ extern "C" {
 
 #[inline]
 pub unsafe fn PySeqIter_Check(op: *mut PyObject) -> c_int {
-    (Py_TYPE(op) == &mut PySeqIter_Type) as c_int
+    (Py_TYPE(op) == addr_of_mut_shim!(PySeqIter_Type)) as c_int
 }
 
 extern "C" {
@@ -19,7 +19,7 @@ extern "C" {
 
 #[inline]
 pub unsafe fn PyCallIter_Check(op: *mut PyObject) -> c_int {
-    (Py_TYPE(op) == &mut PyCallIter_Type) as c_int
+    (Py_TYPE(op) == addr_of_mut_shim!(PyCallIter_Type)) as c_int
 }
 
 extern "C" {