+++ pytests/tests/test_othermod.py
@@ -3,14 +3,14 @@ from hypothesis import strategies as st
 
 from pyo3_pytests import othermod
 
-INTEGER32_ST = st.integers(min_value=(-(2 ** 31)), max_value=(2 ** 31 - 1))
+INTEGER32_ST = st.integers(min_value=(-(2**31)), max_value=(2**31 - 1))
 USIZE_ST = st.integers(min_value=othermod.USIZE_MIN, max_value=othermod.USIZE_MAX)
 
 
 @given(x=INTEGER32_ST)
 def test_double(x):
     expected = x * 2
-    assume(-(2 ** 31) <= expected <= (2 ** 31 - 1))
+    assume(-(2**31) <= expected <= (2**31 - 1))
     assert othermod.double(x) == expected