+++ src/impl_/extract_argument.rs
@@ -103,7 +103,10 @@ pub fn from_py_with_with_default<'py, T>(
 #[cold]
 pub fn argument_extraction_error(py: Python, arg_name: &str, error: PyErr) -> PyErr {
     if error.get_type(py) == PyTypeError::type_object(py) {
-        PyTypeError::new_err(format!("argument '{}': {}", arg_name, error.value(py)))
+        let remapped_error =
+            PyTypeError::new_err(format!("argument '{}': {}", arg_name, error.value(py)));
+        remapped_error.set_cause(py, error.cause(py));
+        remapped_error
     } else {
         error
     }